made deleting temporary files work for lists of files

+++ acstools/acscteforwardmodel.py
@@ -97,10 +97,10 @@ def acscteforwardmodel(input, exec_path='', time_stamps=False, verbose=False,
     # Parse input to get list of filenames to process.
     # acscte.e only takes 'file1,file2,...'
     infiles, dummy_out = parseinput.parseinput(input)
-    
+
     # Make temporary files for input into forward model
     tmp_files = make_tmp_files(infiles)
-    
+
     call_list.append(','.join(tmp_files))
 
     if time_stamps:
@@ -119,13 +119,11 @@ def acscteforwardmodel(input, exec_path='', time_stamps=False, verbose=False,
         call_list.extend(exe_args)
 
     subprocess.check_call(call_list)
-    
+
     # Restore extra FLC extensions to forward model
-    if tmp_files != infiles:
-    
-        restore_files(infiles)
-    
-    
+    restore_files(infiles)
+
+
 def make_tmp_files(infiles):
     """
     If more than 6 extensions, take SCI, ERR, and DQ extensions and make a
@@ -154,6 +152,7 @@ def make_tmp_files(infiles):
 
         outfile = '{}_tmp.fits'.format(file.split('.fits')[0])
 
+        # Update NEXTEND keyword and select extensions for temporary file
         if hdr['NEXTEND'] > 6:
 
             if hdr['SUBARRAY']:
@@ -166,16 +165,18 @@ def make_tmp_files(infiles):
                 new_hdu = hdu[:7]
                 hdr['NEXTEND'] = 6
 
+            # Write temporary file
             new_hdu.writeto(outfile)
 
             tmp_files.append(outfile)
 
+        # If file has typical number of FLCs, output original input
         else:
 
             tmp_files.append(file)
 
     return tmp_files
-    
+
 
 def restore_files(infiles):
     """
@@ -191,35 +192,43 @@ def restore_files(infiles):
 
     from astropy.io import fits
     from shutil import copyfile
+    import glob
 
     for file in infiles:
 
-        copy = '{}_ctefmod.fits'.format(file.split('.fits')[0])
+        base = file.split('.fits')[0]
+        tmp_file = '{}_tmp_ctefmod.fits'.format(base)
+        
+        if os.path.exists(tmp_file):
+        
+            # Make a copy of input file in which to dump forward modeled
+            # extensions
+            copy = '{}_ctefmod.fits'.format(base)
 
-        copyfile(file, copy)
+            copyfile(file, copy)
+            
+            hdu = fits.open(copy, mode='update')
+            tmp_hdu = fits.open(tmp_file)
 
-        tmp_file = '{}_tmp_ctefmod.fits'.format(file.split('.fits')[0])
+            hdr = hdu[0].header
 
-        hdu = fits.open(copy, mode='update')
-        tmp_hdu = fits.open(tmp_file)
+            if hdr['SUBARRAY']:
 
-        hdr = hdu[0].header
+                hdu[1].header = tmp_hdu[1].header
+                hdu[1].data = tmp_hdu[1].data
 
-        if hdr['SUBARRAY']:
+            else:
 
-            hdu[1].header = tmp_hdu[1].header
-            hdu[1].data = tmp_hdu[1].data
+                hdu[1].header = tmp_hdu[1].header
+                hdu[1].data = tmp_hdu[1].data
+                hdu[4].header = tmp_hdu[4].header
+                hdu[4].data = tmp_hdu[4].data
 
-        else:
+            hdu.close()
 
-            hdu[1].header = tmp_hdu[1].header
-            hdu[1].data = tmp_hdu[1].data
-            hdu[4].header = tmp_hdu[4].header
-            hdu[4].data = tmp_hdu[4].data
-
-        hdu.close()
-
-        os.remove('{}_tmp*'.format(file.split('.fits')[0])
+            # Remove temporary files
+            for tmp in glob.glob('{}_tmp*'.format(base)):
+                os.remove(tmp)
 
 
 def getHelpAsString():
